新增一些工具类

+++ androidx/src/main/java/com/androidx/file/FileX.kt
@@ -30,9 +30,10 @@ object FileX {
     }
 
     fun clearCache(context: Context): Long {
-        return deleteDir(context.cacheDir)
+        return cleanDir(context.cacheDir)
     }
 
+    // 会将输入的根目录一起删除掉
     fun deleteDir(dir: File?): Long {
         var deletedSize = 0L
         if (dir == null || !dir.exists()) return 0L
@@ -41,6 +42,25 @@ object FileX {
         for (file in files) {
             deletedSize += if (file.isDirectory) {
                 deleteDir(file)
+            } else {
+                val fileSize = file.length()
+                if (file.delete()) fileSize else 0L
+            }
+        }
+        // 删除自身目录（不计算大小）
+        dir.delete()
+        return deletedSize
+    }
+
+    // 不会删除清理输入的根目录
+    fun cleanDir(dir: File?): Long {
+        var deletedSize = 0L
+        if (dir == null || !dir.exists()) return 0L
+
+        val files = dir.listFiles() ?: return 0L
+        for (file in files) {
+            deletedSize += if (file.isDirectory) {
+                cleanDir(file)
             } else {
                 val fileSize = file.length()
                 if (file.delete()) fileSize else 0L
